feat: implement shared read-only timetable layout for lecturers and technical officers
