+++ TEC-SIS/src/com/fot/system/view/dashboard/lecturer/timetable/LecturerTimetablePanel.java
@@ -1,0 +1,28 @@
+package com.fot.system.view.dashboard.lecturer.timetable;
+import com.fot.system.model.entity.User;
+import com.fot.system.view.dashboard.shared_components.TimeTablePanel;
+
+/**
+ * Provides a lecturer-facing read-only timetable view using the shared timetable layout.
+ * @author janith
+ */
+public class LecturerTimetablePanel extends TimeTablePanel {
+
+    /**
+     * Creates the lecturer timetable panel without edit controls.
+     * @param user logged-in lecturer user
+     * @author janith
+     */
+    public LecturerTimetablePanel(User user) {
+        super();
+    }
+
+    /**
+     * Returns the lecturer timetable subtitle.
+     * @author janith
+     */
+    @Override
+    protected String getPanelSubtitle() {
+        return "View the weekly timetable in a read-only timetable layout with weekdays across the top.";
+    }
+}
